Fixes #1
For each example the plugin file is loaded. When its loaded the first
time Ohai uses the DSL to create a class object that it assigns to a
constant and stores in a particular module. This is usually alright but
does create a problem when you attempt to load this same class definition
over and over again.

To get around that I added to the before/after to capture the current
constants that are defined and remove them after each run. This means
that every specification will now reload the ohai plugin class.

Signed-off-by: Franklin Webber <franklin@chef.io>