Provides support for template ohai plugins
To provide support for template ohai plugins there had to be a change
in how the source is loaded. Files found in the templates path are
assumed to be templates. Templates will have template variables defined
and passed to them. So a new helper `template_variables` has to be
defined.

Signed-off-by: Franklin Webber <franklin@chef.io>